2020.12.17 1.0version

+++ src/main/java/com/example/demo/MyData/Entity/User.java
@@ -6,7 +6,7 @@ import java.util.List;
 @Data
 public class User {
     public int userId;
-    public  int money;
+    public  double money;
     public String username;
     public String userEmail;
